FEATURE created basic debounced search bar and moved onboarding to the MAP UI
