+++ src/pages/Map/Map.tsx
@@ -1,28 +1,31 @@
 import { useState } from 'react';
 import { MapContainer, TileLayer } from 'react-leaflet';
-import mapProviders from '../../utilities/mapProviders.json';
-import UserMarker from '../../components/map_components/UserMarker/UserMarker';
-import Markers from '../../components/map_components/Markers/Markers';
-import Walter from '../../components/UI/Walter/Walter';
-import ListToggle from '../../components/map_UI/ListToggle/ListToggle';
-import ListView from '../../components/map_UI/ListView/ListView';
-import RecenterButton from '../../components/map_UI/RecenterButton/RecenterButton';
-import FiltersDrawer from '../../components/map_UI/FiltersDrawer/FiltersDrawer';
-import MapProviderSelector from '../../components/map_UI/MapProviderSelector/MapProviderSelector';
-import DataFetchingNotifier from '../../components/map_UI/DataFetchingNotifier/DataFetchingNotifier';
+import mapProviders from '@/utilities/mapProviders.json';
+import UserMarker from '@/components/map_components/UserMarker/UserMarker';
+import Markers from '@/components/map_components/Markers/Markers';
+import Walter from '@/components/UI/Walter/Walter';
+import ListToggle from '@/components/map_UI/ListToggle/ListToggle';
+import ListView from '@/components/map_UI/ListView/ListView';
+import RecenterButton from '@/components/map_UI/RecenterButton/RecenterButton';
+import FiltersDrawer from '@/components/map_UI/FiltersDrawer/FiltersDrawer';
+import MapProviderSelector from '@/components/map_UI/MapProviderSelector/MapProviderSelector';
+import DataFetchingNotifier from '@/components/map_UI/DataFetchingNotifier/DataFetchingNotifier';
+import Onboarding from '@/components/map_UI/Onboarding/Onboarding';
+import SearchBar from '@/components/map_UI/SearchBar/SearchBar';
 import { MapSelecter } from './Map.types';
-import Onboarding from '@/components/UI/Onboarding/Onboarding';
 import './leaflet.css'; // to stay up to date, you can always : import 'leaflet/dist/leaflet.css';
+
 
 export default function Map() {
   const [listIsDisplayed, setListIsDisplayed] = useState<boolean>(false);
   const [mapSelecter, setMapSelecter] = useState<MapSelecter>({
     isOpen: false,
-    providerId: Number(localStorage.getItem('mapProviderId')),
+    providerId: Number(localStorage.getItem('mapProviderId')), // TODO convert to ww_map_provider_id
   });
 
   return (
     <>
+      <SearchBar />
       <Onboarding />
       <ListView listState={{ listIsDisplayed, setListIsDisplayed }} />
